deleting confirmation v1

+++ ui/src/App.js
@@ -1,13 +1,11 @@
 import './App.css';
-import {useState} from "react";
-import {useEffect} from 'react';
+import {useEffect, useState} from "react";
 import "milligram";
 import MovieForm from "./MovieForm";
 import MoviesList from "./MoviesList";
 import ActorForm from "./ActorForm";
 import ActorsList from "./ActorsList";
-import { ToastContainer, toast } from 'react-toastify';
-
+import { ToastContentProps, ToastContainer, toast } from 'react-toastify';
 
 function App() {
     const [movies, setMovies] = useState([]);
@@ -71,21 +69,41 @@ function App() {
         const response = await fetch(`/movies/${movie.id}`, {
             method: 'DELETE',
         });
-        if (response.ok) {
+        if (response.ok) {toast.success ("Movie deleted successfully")
             const nextMovies = movies.filter(m => m !== movie);
             setMovies(nextMovies);
         }
+        else {toast.error ("Failed to delete movie")}
     }
 
     async function handleDeleteActor(actor) {
-        const response = await fetch(`/actors/${actor.id}`, {
-            method: 'DELETE',
+        toast(deleteConfirm, {
+            onClose(reason){
+                if (reason === "delete"){
+                    fetch(`/actors/${actor.id}`, {
+                      method: 'DELETE',
+                    }).then(response => {
+                        if (response.ok) {toast.success ("Actor deleted successfully")
+                            const nextActors = actors.filter(m => m !== actor);
+                            setActors(nextActors);
+                        }
+                        else {toast.error ("Failed to delete actor")}
+                    })
+                }
+            }
         });
-        if (response.ok) {
-            const nextActors = actors.filter(m => m !== actor);
-            setActors(nextActors);
-        }
     }
+
+    function deleteConfirm({ closeToast }) {
+      return (
+        <div class="confirmation-toast">
+          Are you sure you want to delete this?
+          <button class="button" onClick={() => closeToast("delete")}>Yes, delete</button>
+          <button class="button button-outline" onClick={() => closeToast("cancel")}>Cancel</button>
+        </div>
+      )
+    }
+
 
     return (
         <div className="container">
